django-token-mw/requirements.txt
django==4
pyjwt==2.4.0
python-jose[cryptography]==3.3.0
python-keycloak==1.6.0
